+++ src/components/TextIcon.svelte
